Make Stats a public page
