+++ app/assets/stylesheets/stats.scss
