+++ app/views/administrations/index.html.haml
@@ -7,9 +7,6 @@
 
         = f.submit 'Créer un administrateur', class: 'btn btn-success', id: 'submit_new_administrateur'
 
-    .text-center
-      = link_to 'Stats', administrations_stats_path, style: 'margin-bottom: 50px; display: block;', 'data-no-turbolink': true
-
     = smart_listing_render :admins
 
     %br
