+++ app/views/stats/index.html.haml
